change format, added style for component Button
